style: separete others routes
