+++ backend/kemomimi-api/src/main.rs
@@ -1,29 +1,7 @@
 use axum::{extract::Host, http::Method, Router};
 use axum_extra::extract::CookieJar;
 use chrono::Utc;
-use openapi::{
-    apis::{
-        categories::{
-            Categories, CategoriesCategoryIdDeleteResponse, CategoriesCategoryIdGetResponse,
-            CategoriesCategoryIdPutResponse, CategoriesGetResponse, CategoriesPostResponse,
-        },
-        private_items::{
-            PrivateItems, PrivateItemsGetResponse, PrivateItemsPostResponse,
-            PrivateItemsPrivateItemIdDeleteResponse, PrivateItemsPrivateItemIdGetResponse,
-            PrivateItemsPrivateItemIdPutResponse,
-        },
-        products::{
-            Products, ProductsGetResponse, ProductsPostResponse, ProductsProductIdDeleteResponse,
-            ProductsProductIdGetResponse, ProductsProductIdPutResponse,
-        },
-        users::{
-            Users, UsersGetResponse, UsersPostResponse, UsersUserIdDeleteResponse,
-            UsersUserIdGetResponse, UsersUserIdPutResponse,
-        },
-    },
-    models::{self},
-    server::new,
-};
+use openapi::server::new;
 use sqlx::{PgPool, Pool, Postgres};
 use std::env;
 use std::sync::Arc;
@@ -59,593 +37,5 @@ async fn main() {
 impl AsRef<ApiImpl> for ApiImpl {
     fn as_ref(&self) -> &ApiImpl {
         self
-    }
-}
-
-impl Categories for ApiImpl {
-    #[doc = " カテゴリ削除."]
-    #[doc = ""]
-    #[doc = " CategoriesCategoryIdDelete - DELETE /categories/{category-id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn categories_category_id_delete<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::CategoriesCategoryIdDeletePathParams,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<CategoriesCategoryIdDeleteResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " カテゴリ情報取得."]
-    #[doc = ""]
-    #[doc = " CategoriesCategoryIdGet - GET /categories/{category-id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn categories_category_id_get<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::CategoriesCategoryIdGetPathParams,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<CategoriesCategoryIdGetResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " カテゴリ情報更新."]
-    #[doc = ""]
-    #[doc = " CategoriesCategoryIdPut - PUT /categories/{category-id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn categories_category_id_put<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::CategoriesCategoryIdPutPathParams,
-        body: models::Category,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<CategoriesCategoryIdPutResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " カテゴリ一覧取得."]
-    #[doc = ""]
-    #[doc = " CategoriesGet - GET /categories"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn categories_get<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<CategoriesGetResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " カテゴリ新規登録."]
-    #[doc = ""]
-    #[doc = " CategoriesPost - POST /categories"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn categories_post<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        body: models::Category,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<CategoriesPostResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-}
-
-impl PrivateItems for ApiImpl {
-    #[doc = " 私物一覧取得."]
-    #[doc = ""]
-    #[doc = " PrivateItemsGet - GET /private-items"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn private_items_get<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<PrivateItemsGetResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 私物新規登録."]
-    #[doc = ""]
-    #[doc = " PrivateItemsPost - POST /private-items"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn private_items_post<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        body: models::PrivateItem,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<PrivateItemsPostResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 私物削除."]
-    #[doc = ""]
-    #[doc = " PrivateItemsPrivateItemIdDelete - DELETE /private-items/{private-item-id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn private_items_private_item_id_delete<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::PrivateItemsPrivateItemIdDeletePathParams,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<PrivateItemsPrivateItemIdDeleteResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 私物情報取得."]
-    #[doc = ""]
-    #[doc = " PrivateItemsPrivateItemIdGet - GET /private-items/{private-item-id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn private_items_private_item_id_get<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::PrivateItemsPrivateItemIdGetPathParams,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<PrivateItemsPrivateItemIdGetResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 私物情報更新."]
-    #[doc = ""]
-    #[doc = " PrivateItemsPrivateItemIdPut - PUT /private-items/{private-item-id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn private_items_private_item_id_put<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::PrivateItemsPrivateItemIdPutPathParams,
-        body: models::PrivateItem,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<PrivateItemsPrivateItemIdPutResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-}
-
-impl Products for ApiImpl {
-    #[doc = " 製品一覧取得."]
-    #[doc = ""]
-    #[doc = " ProductsGet - GET /products"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn products_get<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<ProductsGetResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 製品新規登録."]
-    #[doc = ""]
-    #[doc = " ProductsPost - POST /products"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn products_post<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        body: models::Product,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<ProductsPostResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 製品削除."]
-    #[doc = ""]
-    #[doc = " ProductsProductIdDelete - DELETE /products/{product-id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn products_product_id_delete<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::ProductsProductIdDeletePathParams,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<ProductsProductIdDeleteResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 製品情報取得."]
-    #[doc = ""]
-    #[doc = " ProductsProductIdGet - GET /products/{product-id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn products_product_id_get<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::ProductsProductIdGetPathParams,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<ProductsProductIdGetResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 製品情報更新."]
-    #[doc = ""]
-    #[doc = " ProductsProductIdPut - PUT /products/{product-id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn products_product_id_put<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::ProductsProductIdPutPathParams,
-        body: models::Product,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<ProductsProductIdPutResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-}
-
-impl Users for ApiImpl {
-    #[doc = " 部員一覧取得."]
-    #[doc = ""]
-    #[doc = " UsersGet - GET /users"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn users_get<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<UsersGetResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 部員新規登録."]
-    #[doc = ""]
-    #[doc = " UsersPost - POST /users"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn users_post<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        body: models::User,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<UsersPostResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 部員削除."]
-    #[doc = ""]
-    #[doc = " UsersUserIdDelete - DELETE /users-{user_id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn users_user_id_delete<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::UsersUserIdDeletePathParams,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<UsersUserIdDeleteResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 部員情報取得."]
-    #[doc = ""]
-    #[doc = " UsersUserIdGet - GET /users-{user_id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn users_user_id_get<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::UsersUserIdGetPathParams,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<UsersUserIdGetResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
-    }
-
-    #[doc = " 部員情報更新."]
-    #[doc = ""]
-    #[doc = " UsersUserIdPut - PUT /users-{user_id}"]
-    #[must_use]
-    #[allow(
-        elided_named_lifetimes,
-        clippy::type_complexity,
-        clippy::type_repetition_in_bounds
-    )]
-    fn users_user_id_put<'life0, 'async_trait>(
-        &'life0 self,
-        method: Method,
-        host: Host,
-        cookies: CookieJar,
-        path_params: models::UsersUserIdPutPathParams,
-        body: models::User,
-    ) -> ::core::pin::Pin<
-        Box<
-            dyn ::core::future::Future<Output = Result<UsersUserIdPutResponse, ()>>
-                + ::core::marker::Send
-                + 'async_trait,
-        >,
-    >
-    where
-        'life0: 'async_trait,
-        Self: 'async_trait,
-    {
-        todo!()
     }
 }
