+++ backend/kemomimi-api/src/routes.rs
@@ -1,1 +1,5 @@
+pub mod categories;
+pub mod private_items;
+pub mod products;
 pub mod public_items;
+pub mod users;
