+++ backend/kemomimi-api/src/routes/categories.rs
@@ -1,0 +1,1 @@
+pub mod index;
