+++ backend/kemomimi-api/src/routes/private_items.rs
@@ -1,0 +1,1 @@
+pub mod index;
