+++ backend/kemomimi-api/src/routes/products.rs
@@ -1,0 +1,1 @@
+pub mod index;
