+++ backend/kemomimi-api/src/routes/users.rs
@@ -1,0 +1,1 @@
+pub mod index;
